requirements.txt
google-genai>=0.2.0
streamlit>=1.36.0
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
